+++ api/src/main/resources/mybatis-mapper/TradePostMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.hcs.mapper.tradePost.TradePostMapper">
+<mapper namespace="com.hcs.mapper.TradePostMapper">
 
     <select id="findByTitle" parameterType="String" resultType="com.hcs.domain.TradePost">
         select *
