+++ api/src/main/resources/mybatis-mapper/UserMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.hcs.mapper.user.UserMapper">
+<mapper namespace="com.hcs.mapper.UserMapper">
 
     <select id="findByEmail" parameterType="String" resultType="com.hcs.domain.User">
         select *
