+++ api/src/test/java/com/hcs/controller/UserControllerTest.java
@@ -1,10 +1,10 @@
-package com.hcs.user;
+package com.hcs.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hcs.config.EnableMockMvc;
 import com.hcs.domain.User;
 import com.hcs.dto.SignUpDto;
-import com.hcs.mapper.user.UserMapper;
+import com.hcs.mapper.UserMapper;
 import com.ulisesbocchio.jasyptspringboot.annotation.EnableEncryptableProperties;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
