+++ api/src/test/java/com/hcs/mapper/ClubMapperTest.java
@@ -2,7 +2,6 @@ package com.hcs.mapper;
 
 import com.hcs.domain.Club;
 import com.hcs.domain.User;
-import com.hcs.mapper.user.UserMapper;
 import com.ulisesbocchio.jasyptspringboot.annotation.EnableEncryptableProperties;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
